rag_backup/requirements.txt


fastapi
uvicorn[standard]

google-generativeai
python-dotenv
faiss-cpu


PyMuPDF
docx2txt

google-api-python-client